Refactor app.js to import and use separate route files for auth, messages, and home
